More PABotBase2.

+++ Common/PABotBase2/ConnectionLayer/PABotBase2_ConnectionDebug.cpp
@@ -115,7 +115,7 @@ void PacketSender::print(bool ascii) const{
     std::cout << "Buffer Head:       " << m_buffer_head << std::endl;
     std::cout << "Buffer Tail:       " << m_buffer_tail << std::endl;
     std::cout << "Stream Offset:     " << m_stream_offset << std::endl;
-    std::cout << "Retransmit Seqnum: " << m_retransmit_seqnum << std::endl;
+    std::cout << "Retransmit Seqnum: " << (int)m_retransmit_seqnum << std::endl;
     for (uint8_t seqnum = m_slot_head; seqnum != m_slot_tail; seqnum++){
         size_t offset = ~m_offsets[seqnum & SLOTS_MASK];
         std::cout << "Offset: " << offset << std::endl;
